+++ rustsec/src/osv/ranges_for_advisory.rs
